Organized Javascript directories
