+++ scripts/main.js
@@ -1,7 +1,6 @@
 import initializeDetailButtonEvents from './dialogs.js'
-import { useFish } from './FishDataProvider.js'
-import FishList from './FishList.js'
-
+import { useFish } from './fish/FishDataProvider.js'
+import FishList from './fish/FishList.js'
 // initializeDetailButtonEvents()
 
 const arrayOfTheFishObjects = useFish()
